requirements.txt
alembic==1.7.5
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
bidict==0.22.0
billiard==3.6.4.0
blinker==1.4
celery==5.2.7
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
dnspython==2.2.1
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Dance==5.1.0
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SocketIO==5.3.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
kombu==5.2.4
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
oauthlib==3.1.1
packaging==21.3
prompt-toolkit==3.0.31
psycopg2==2.9.3
pyparsing==3.0.9
python-dotenv==0.19.2
python-engineio==4.3.4
python-socketio==5.7.1
pytz==2021.3
redis==4.3.4
requests==2.26.0
requests-oauthlib==1.3.0
simple-websocket==0.8.1
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.29
urllib3==1.26.7
URLObject==2.4.3
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.14.1
wsproto==1.2.0
zope.event==4.5.0
zope.interface==5.4.0
